requirements.txt
appdirs==1.4.3
asgiref==3.3.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
channels==3.0.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.2.1
daphne==3.0.1
distlib==0.3.0
distro==1.4.0
Django==3.1.3
django-crispy-forms==1.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
html5lib==1.0.1
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
mysqlclient==2.0.1
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.4
redis==3.5.3
requests==2.22.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
webencodings==0.5.1
zope.interface==5.2.0
